==> 0B_lunDate_api.ipynb <==
from urllib.request import urlopen
from urllib.request import Request 
from urllib.parse import urlencode, quote_plus

import urllib.request as ul
import xmltodict
import json
import sys
import io
import pandas as pd
import requests

### 연월일 분리위한 데이터 프레임 만들기

start = pd.to_datetime('20130101')
end = pd.to_datetime('20190331') 
df = pd.DataFrame(columns=['date'])
df['date'] = pd.date_range(start,end,freq='D')

df['Year'] = df['date'].dt.year 
df['Month'] = df['date'].dt.month 
df['Day'] = df['date'].dt.day

df

### 월,일의 한자리(ex. 1,2,3...) 앞에 0붙여야함
def convert_month(df):
    for i in range(len(df)):
        month = df['Month'].iloc[i]
        if int(month) < 10:
            month = "0"+str(month)
            df['Month'].iloc[i] = month
            
def convert_day(df):
    for i in range(len(df)):
        day = df['Day'].iloc[i]
        if int(day) < 10:
            day = "0"+str(day)
            df['Day'].iloc[i] = day        

convert_month(df)
convert_day(df)
df

### API로 음력 받아오기

data = pd.DataFrame()


for i in range(len(df)):
    
    Year = df['Year'].iloc[i]
    Month = df['Month'].iloc[i]
    Day = df['Day'].iloc[i]
               
    url = 'http://apis.data.go.kr/B090041/openapi/service/LrsrCldInfoService/getLunCalInfo?serviceKey=개인정보모자이크&solYear='+str(Year)+'&solMonth='+str(Month)+'&solDay='+str(Day)

    requestd = Request(url)
    requestd.get_method = lambda: 'GET'
    response_body = urlopen(requestd).read()
    resp = xmltodict.parse(response_body)
    resp_dic = json.loads(json.dumps(resp))
    #print(resp_dic)
    #print(i)

    data = data.append(resp_dic['response']['body']['items']['item'],ignore_index=True)

print("********end********")

raw_api = data[['lunYear','lunMonth','lunDay']]
raw_api

### 만들어둔 데이터 프레임이랑 합치기

lun_2013 = pd.concat([df,raw_api],axis=1) 
lun_2013

lun_2013.to_csv("./make_data/0B_음력날짜.csv",index=False) ##여기까지 20130101~20190331

### 데이터 프레임 만들기 (19881001~20121231)

start = pd.to_datetime('19881001')
end = pd.to_datetime('20121231') 
df2 = pd.DataFrame(columns=['date'])
df2['date'] = pd.date_range(start,end,freq='D')

df2['Year'] = df2['date'].dt.year 
df2['Month'] = df2['date'].dt.month 
df2['Day'] = df2['date'].dt.day

df2

convert_month(df2)
convert_day(df2)
df2

### API 받아오기 (19881001~20121231)

data = pd.DataFrame()


for i in range(len(df2)):
    
    Year = df2['Year'].iloc[i]
    Month = df2['Month'].iloc[i]
    Day = df2['Day'].iloc[i]
               
    url = 'http://apis.data.go.kr/B090041/openapi/service/LrsrCldInfoService/getLunCalInfo?serviceKey=개인정보모자이크&solYear='+str(Year)+'&solMonth='+str(Month)+'&solDay='+str(Day)

    requestd = Request(url)
    requestd.get_method = lambda: 'GET'
    response_body = urlopen(requestd).read()
    resp = xmltodict.parse(response_body)
    resp_dic = json.loads(json.dumps(resp))
    #print(resp_dic)
    #print(i)

    data = data.append(resp_dic['response']['body']['items']['item'],ignore_index=True)

print("********end********")

raw_api2 = data[['lunYear','lunMonth','lunDay']]

lun_1988 = pd.concat([df2,raw_api2],axis=1) #열방향으로 음력합침
lun_1988

lun_final_1988 = pd.concat([lun_1988,lun_2013],axis=0) #행방향으로 날짜합침

lun_final_1988

lun_final_1988.reset_index(drop=True, inplace=True)
lun_final_1988

lun_final_1988.to_csv("./make_data/0B_1988_음력날짜.csv",index=False)
