==> 0C_sun_api.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from glob import glob
import warnings
warnings.filterwarnings('ignore')

from urllib.request import urlopen
from urllib.request import Request 
from urllib.parse import urlencode, quote_plus

import urllib.request as ul
import xmltodict
import json
import sys
import io

import requests

from datetime import datetime
start = pd.to_datetime('20130101')
end=pd.to_datetime('20190331')
df=pd.DataFrame(columns=['연월일'])
df['연월일']=pd.date_range(start, end,freq='D')

df

def make_locdate(x):
    Y, M, D = x.split('-')
    return Y+M+D

df['연월일']=df['연월일'].astype(str)
df.info()

df['locdate'] = df['연월일'].apply(lambda x:make_locdate(x))
df

url = 'http://apis.data.go.kr/B090041/openapi/service/RiseSetInfoService/getAreaRiseSetInfo'

params ={'serviceKey' : '0GE5dffPQ2w31OZOACOdSgFRQzPihrAaslHTL1WR8UajH2KJivTGQfGkmw8pCFaWIr9sG2Vh4YEuaRk4rvCgSg==', 'locdate' : '20220505', 'location' : '서울' }
response = requests.get(url, params)
resp = xmltodict.parse(response.text)
resp

url = 'http://apis.data.go.kr/B090041/openapi/service/RiseSetInfoService/getAreaRiseSetInfo'

for i in range(0, df.shape[0]):
    locdate = df.loc[i,'locdate']
    params ={'serviceKey' : '개인정보모자이크==', 'locdate' : locdate, 'location' : '서울' }
    response = requests.get(url, params)
    resp = xmltodict.parse(response.text)
    print(resp)
    df.loc[i, 'sunrise'] = resp['response']['body']['items']['item']['sunrise']
    df.loc[i, 'sunset'] = resp['response']['body']['items']['item']['sunset']
    print(i)
print('end')

df

df['sunset'] = df['sunset'].astype(int)//100
df['sunrise']= df['sunrise'].astype(int)//100

df['일사']= df["sunset"] - df["sunrise"]

df.rename(columns={'연월일':'date'}, inplace=True)

df2 = df[['date','sunrise','sunset','일사']]
df2

### 1. train 데이터(2013-01-01 ~ 2018-12-31)

start = df2['date']>='2013-01-01'
end = df2['date']<='2018-12-31'
train_일출일몰 = df2[start & end]
train_일출일몰

train_일출일몰.to_csv('./make_data/0C_train_일출일몰_예보.csv', index=False)

### 2. test 데이터(2019-01-01 ~ 2019-03-31)

start2 = df2['date']>='2019-01-01'
end2 = df2['date']<='2019-03-31'
test_일출일몰 = df2[start2 & end2]
test_일출일몰

test_일출일몰.to_csv('./make_data/0C_test_일출일몰_예보.csv', index=False)
