==> 0D_temp_30_years.ipynb <==
import pandas as pd
import numpy as np
from glob import glob
import warnings
warnings.filterwarnings('ignore')
import io
from sklearn.metrics import mean_absolute_error

## 1. 기상청 단기예보+중기예보 가져와서 10일치 예보 만들기 

data_year_temperature = pd.read_csv('./data/청운효자동_3시간기온_201812_201812.csv')  # 3시간 기온데이타
data_year = data_year_temperature[[' format: day', 'hour', 'forecast']]   # 하나의 데이타프레임에 기온데이타 만들기 
data_year['Temperature'] = data_year_temperature[data_year_temperature.columns[-1]]  

year =data_year_temperature.columns[3][31:35]  # year는 2018
month = data_year_temperature.columns[3][35:37] #month는 12

date = f'{year}-{month}-' +data_year[' format: day'].astype(str).str.split(' ').str[-1] + ' ' + (data_year['hour'].astype(int)//100).astype(str) + ':00' # date에 년월일데이타 만들기 
date = pd.to_datetime(date) + pd.DateOffset(hours=9) 

data_fcst = pd.DataFrame(columns=['Forecast time', 'forecast', 'Temperature'                       
                                  ])   # DataFrame 만들어 
data_fcst['Forecast time'] = date # date 채우기  
data_fcst[['forecast', 'Temperature']] = data_year[['forecast', 'Temperature'                                                               
                                                                 ]]

# Forecast time에 forecast 시간 더해서 Forecast_time 만들기 
fcst_data = data_fcst.copy() 

forecastchange=[]
forecast = fcst_data['forecast'].tolist() 
forecastTime= fcst_data['Forecast time'].tolist()
for f,t in zip(forecast,forecastTime):
    forecastchange.append(t + pd.DateOffset(hours=f))     
fcst_data['Forecast_time'] = forecastchange 

fcst_data.isnull().sum()

# Forecast_time이 2019년인것만 가져오고 duplicated drop
fcst_data = fcst_data[['Forecast_time','Temperature']]
fcst_data = fcst_data.drop_duplicates(subset=['Forecast_time'], keep='last').reset_index(drop=True)
fcst_data

#날짜 DataFrame만들기 
date = pd.DataFrame(pd.date_range(start='2019-01-01', end='2019-01-04', freq='H', name='Forecast_time', closed='left'))

# fcst_data와 합쳐서 보간하기 
fcst_data_ = pd.merge(fcst_data,date, how='outer', on='Forecast_time')
fcst_data_ = fcst_data_.sort_values(by='Forecast_time').reset_index(drop=True)

fcst_data__ = fcst_data_[243:].reset_index(drop=True) # 1월만 자르기 

fcst_data_inter = fcst_data__.interpolate()

# 중기예보 
fcst_midterm = pd.read_csv('./data/FCT_MEDM_20211203165632.csv', encoding='CP949')
fcst_midterm = fcst_midterm[(fcst_midterm['도시']=='서울') & (fcst_midterm['발표시각'].str.contains('2018-12-31 18시'))] # 서울/오후6시예보만 가져와서 
#날짜 dataframe만들기 
date2 = pd.DataFrame( pd.date_range(start='2019-01-03', end='2019-01-11', freq='H',name='Forecast_time' , closed='left'))

# 중기예보의 최저기온,최고기온 리스트로 만들어 date2 'Temperature' 열 ## 오전6시에 최저기온, 오후3시에 최고기온으로 채우기 (바꿔가며해봄)
low = fcst_midterm['최저기온'].tolist()
high = fcst_midterm['최고기온'].tolist()
date2['Temperature']=np.NAN

# 최저기온 채우기 
for i,j in enumerate(low): 
    date2.loc[24*i+6,'Temperature'] = j 

#최고기온 채우기 
for i,j in enumerate(high):
    date2.loc[24*i+15,'Temperature'] = j 

date2 = date2.drop(date2.index[0:23],axis=0) # 1/3일은 단기예보와 겹치므로 drop 

date2

# 단기예보와 합쳐서 보간하기
temp_for10 = pd.concat([fcst_data_inter,date2],axis=0)

temp_for10['Temperature']= temp_for10['Temperature'].astype(float)
temp_for10_inter = temp_for10.interpolate()
temp_for10_inter = temp_for10_inter.drop_duplicates(subset=['Forecast_time'])

temp_for10_inter.to_csv('./make_data/0D_weather_10_float.csv', index=False) # 단기예보 3일치+중기예보 7일치 합쳐서 10일 예보만듬. 

##  2. 관측치 30년치에서 여러가지 평균내보고 10일 예보와 합쳐 관측치와 오차점수 보기 

obs_data_path = './data/temp_30years' # 30년치 1988부터 
obs_data_list = sorted(glob(obs_data_path+'/*.csv')) # 기상청 관측 (10년치)  

obs_data_list #2008년부터 2018년까지 (음력날짜반영위해 11년) 
# 30년, 1988년부터 2018년까지 

df_merge = pd.DataFrame() # 데이터 합치기 
for idx,filename in enumerate(obs_data_list):
    df = pd.read_csv(filename, encoding='CP949',usecols=['일시','기온(°C)'])
    df_merge = df_merge.append(df,ignore_index=True)

df_merge = df_merge.sort_values('일시').reset_index(drop=True) 

df_merge_date = pd.DataFrame(pd.date_range(start='1988-01-01', end='2019-01-01',freq = 'h',name='일시', closed='left')) #30년치 날짜 만들기  
df_merge_date['일시'] = df_merge_date['일시'].astype(str).str[:16]
print(df_merge_date.shape) 

df_date = pd.merge(df_merge_date, df_merge,how='left',on='일시')
df_date

train_temp = df_date[219168:].reset_index(drop=True)
train_temp.columns=['date','temp']
train_temp

train_temp.isnull().sum()

train_temp = train_temp.interpolate()

train_temp.to_csv('./make_data/0D_train_temp_서울.csv',index=False) # 훈련데이터:사용 

# 온도데이터와 날짜데이터 병합 하고 결측치 채우기 
df_merged = pd.merge(df_merge, df_merge_date, how='outer', on='일시')
print(df_merged[df_merged['기온(°C)'].isnull()]) #결측치 확인 8391개 
df_merged = df_merged.sort_values('일시').reset_index(drop=True)  

df_merged = df_merged.interpolate() # 보간하기 

lunar_date = pd.read_csv('./make_data/0B_1988_음력날짜.csv')  ### 2019년 1/1~3/31일까지는 음력으로 11/26~2/25  

lunar_date

df_merged['date']= df_merged['일시'].str[:10] #10년 데이터에서 date날짜 만들기 
df_merged_lunar = pd.merge(df_merged,lunar_date, how='inner', on='date') # 10년 기온데이터와 음력날짜 병합 
df_merged_lunar = df_merged_lunar.sort_values('일시').reset_index(drop=True)
df_merged_lunar['time'] = df_merged_lunar['일시'].str[11:13].astype(int)

df_merged_lunar

df_merged_lunar['date'] = pd.to_datetime(df_merged_lunar['date'], format='%Y-%m-%d')
df_merged_lunar['time'] = df_merged_lunar['time']+1
df_merged_lunar['temp'] = df_merged_lunar['기온(°C)']
df_merged_lunar['year'] = df_merged_lunar['Year']
df_merged_lunar['month'] = df_merged_lunar['Month']
df_merged_lunar['day'] = df_merged_lunar['Day']
df_merged_lunar['weekday'] = df_merged_lunar['date'].dt.weekday    

columns = ['일시','date','time','temp','year','month','day','weekday','lunYear','lunMonth','lunDay']

df_merged_lunar = df_merged_lunar[columns]

df_merged_lunar

df_merged_lunar.to_csv('./make_data/0D_temp_1988_서울.csv', index=False)
