==> 3_model_공급사별_제출용.ipynb <==
import pandas as pd
import numpy as np
import seaborn as sns
import lightgbm as lgb
import xgboost as xgb
from sklearn.ensemble import ExtraTreesRegressor
from sklearn.ensemble import RandomForestRegressor
import joblib
import copy
import warnings
warnings.filterwarnings('ignore')

total = pd.read_csv('./make_data/1_final_train.csv')
test = pd.read_csv('./make_data/2_final_test.csv') 
submission = pd.read_csv('data/sample_submission.csv')
weather_10 = pd.read_csv("./make_data/0D_weather_10_float.csv")
ensemble = copy.deepcopy(submission)

def gas_nmae(true_df, pred_df):
    true = true_df.iloc[:,0].to_numpy()
    pred = pred_df.iloc[:,0].to_numpy()
    score = np.mean((np.abs(true-pred))/true)
    return score

### 1,2,3월만 학습에 사용하기 위해 추출

total = total[(total.month==1)|(total.month==2)|(total.month==3)]

### 업체별 train, test 세트를 리스트로 만들기

train_list=[]
for chain in range(7):
    train0 = total[total.s_chain==chain]
    test0 = test[test.s_chain==chain]
    tlist = [train0, test0]
    train_list.append(tlist)    

### LGBM 모델

train_years = [2013,2014,2015,2016,2017]
val_years = [2018]
idx = 0
sco_list = []

for train_df, test_df in train_list:
    
    train = train_df[train_df['year'].isin(train_years)]
    val = train_df[train_df['year'].isin(val_years)]

    features = ['time','weekday', 'no_working', 'temp', 'lunMonth', 'lunDay', 'month', '일사']
    
    train_x = train[features]
    train_y = train['s_qty']
    val_x = val[features]
    val_y = val['s_qty']
    test_x = test_df[features]

    train_x = train_x.reset_index(drop=True)
    train_y = train_y.reset_index(drop=True)
    val_x = val_x.reset_index(drop=True)
    val_y = val_y.reset_index(drop=True)
    test_x = test_x.reset_index(drop=True)
    
    d_train = lgb.Dataset(train_x, train_y)
    d_val = lgb.Dataset(val_x, val_y)

    params = {
        'objective': 'regression',
        'metric':'mae',
        'max_depth' : 10,
        'seed':42
    }

    model = lgb.train(params, d_train, 500, d_val, verbose_eval=50, early_stopping_rounds=10)
    joblib.dump(model, f'./models/lgbm_model_{idx}.pkl')
    pred2 = model.predict(val_x)
    pred_df = pd.DataFrame(pred2)
    valy_df = pd.DataFrame(val_y)
    nmae_score = gas_nmae(valy_df, pred_df)
    sco_list.append(nmae_score)
    preds = model.predict(test_x)
    submission.iloc[0+(2160*idx):2160+(2160*idx), 1] = preds
    idx += 1
    
print(sco_list)
print('score =', np.array(sco_list).mean())

ensemble['lgbm'] = submission['공급량']

### XGB 모델

train_years = [2013,2014,2015,2016,2017]
val_years = [2018]
idx = 0
sco_list = []

for train_df, test_df in train_list:
    
    train = train_df[train_df['year'].isin(train_years)]
    val = train_df[train_df['year'].isin(val_years)]

    features = ['time','weekday', 'no_working', 'temp', 'lunMonth', 'lunDay', 'month', '일사']

    train_x = train[features]
    train_y = train['s_qty']
    val_x = val[features]
    val_y = val['s_qty']
    test_x = test_df[features]

    train_x = train_x.reset_index(drop=True)
    train_y = train_y.reset_index(drop=True)
    val_x = val_x.reset_index(drop=True)
    val_y = val_y.reset_index(drop=True)
    test_x = test_x.reset_index(drop=True)

    dtrain = xgb.DMatrix(data=train_x, label = train_y)
    dval = xgb.DMatrix(data=val_x, label = val_y)
    wlist = [(dtrain, 'train'), (dval,'eval')]

    params = {
        'learning_rate': 0.05, 
        'objective': 'reg:squarederror',
        'metric':'mae', 
        'seed':42
    }

    model = xgb.train( params, dtrain, 2000, evals=wlist, verbose_eval=200, early_stopping_rounds=100)
    joblib.dump(model, f'./models/xgb_model_{idx}.pkl')
    val_x2 = xgb.DMatrix(val_x)
    pred2 = model.predict(val_x2)
    pred_df = pd.DataFrame(pred2)
    valy_df = pd.DataFrame(val_y)
    nmae_score = gas_nmae(valy_df, pred_df)
    sco_list.append(nmae_score)
    test_x = xgb.DMatrix(test_x)
    preds = model.predict(test_x)
    submission.iloc[0+(2160*idx):2160+(2160*idx), 1] = preds
    idx += 1
    
print(sco_list)
print(np.array(sco_list).mean())

ensemble['xgb'] = submission['공급량']

### ExtraTree 모델

train_years = [2013,2014,2015,2016,2017]
val_years = [2018]
idx = 0
sco_list = []

for train_df, test_df in train_list:
    
    train = train_df[train_df['year'].isin(train_years)]
    val = train_df[train_df['year'].isin(val_years)]

    features = ['time','weekday', 'no_working', 'temp', 'lunMonth', 'lunDay', 'month', '일사']

    train_x = train[features]
    train_y = train['s_qty']
    val_x = val[features]
    val_y = val['s_qty']
    test_x = test_df[features]

    train_x = train_x.reset_index(drop=True)
    train_y = train_y.reset_index(drop=True)
    val_x = val_x.reset_index(drop=True)
    val_y = val_y.reset_index(drop=True)
    test_x = test_x.reset_index(drop=True)

    model = ExtraTreesRegressor(max_depth=12, n_estimators=130)
    model.fit(train_x, train_y)
    joblib.dump(model, f'./models/ET_model_{idx}.pkl')
    pred2 = model.predict(val_x)

    pred_df = pd.DataFrame(pred2)
    valy_df = pd.DataFrame(val_y)
    nmae_score = gas_nmae(valy_df, pred_df)
    sco_list.append(nmae_score)
    preds = model.predict(test_x)
    print(type(preds))
    submission.iloc[0+(2160*idx):2160+(2160*idx), 1] = preds
    idx += 1
    
print(sco_list)
print(np.array(sco_list).mean())

ensemble['extree'] = submission['공급량']

### RandomForest 모델

train_years = [2013,2014,2015,2016,2017]
val_years = [2018]
idx = 0
sco_list = []

for train_df, test_df in train_list:
    
    train = train_df[train_df['year'].isin(train_years)]
    val = train_df[train_df['year'].isin(val_years)]

    features = ['time','weekday', 'no_working', 'temp', 'lunMonth', 'lunDay', 'month', '일사']

    train_x = train[features]
    train_y = train['s_qty']
    val_x = val[features]
    val_y = val['s_qty']
    test_x = test_df[features]

    train_x = train_x.reset_index(drop=True)
    train_y = train_y.reset_index(drop=True)
    val_x = val_x.reset_index(drop=True)
    val_y = val_y.reset_index(drop=True)
    test_x = test_x.reset_index(drop=True)
    
    model = RandomForestRegressor(max_depth=12, n_estimators=100, random_state=0)
    model.fit(train_x, train_y)
    joblib.dump(model, f'./models/rf_model_{idx}.pkl')
    pred2 = model.predict(val_x)

    pred_df = pd.DataFrame(pred2)
    valy_df = pd.DataFrame(val_y)
    nmae_score = gas_nmae(valy_df, pred_df)
    sco_list.append(nmae_score)
    preds = model.predict(test_x)
    submission.iloc[0+(2160*idx):2160+(2160*idx), 1] = preds
    idx += 1
    
print(sco_list)
print(np.array(sco_list).mean())

ensemble['rf'] = submission['공급량']

### 모델 결과 조합

submission['공급량'] = (ensemble['extree'] + ensemble['rf'] + ensemble['lgbm'] + ensemble['xgb']) / 4

submission

submission.to_csv('./submit/3_submission_ensemble.csv', index=False)
